zapcore: break out sampler module
